Generalize the npm package into a full lilbee launcher
The package is now named lilbee and its main bin passes any command
through to the standalone binary: npm install -g lilbee gives the whole
CLI (chat TUI, serve, model management), bootstrapping the binary on
first use. The mcp route keeps the launcher-level behaviors: the
remote bridge when LILBEE_URL is set, --data-dir/env passthrough, and
the prepare subcommand. lilbee-mcp stays as a second bin so MCP host
configs stay one token.
