+++ packages/lilbee/bin/lilbee.mjs
@@ -1,0 +1,3 @@
+#!/usr/bin/env node
+import { runAndReport } from "../lib/cli.mjs";
+runAndReport(process.argv.slice(2));
