+++ packages/lilbee/lib/assets.mjs
