+++ packages/lilbee/lib/cli.mjs
@@ -1,0 +1,116 @@
+/**
+ * Launcher entry logic shared by both bins.
+ *
+ * `lilbee <argv>` routes: `prepare` (download only), `mcp` (stdio MCP server,
+ * or the remote bridge when LILBEE_URL is set), everything else passes through
+ * to the real binary verbatim. All launcher output goes to stderr; stdout
+ * belongs to the command (for `mcp`, it is the MCP wire).
+ */
+
+import { spawn, execFileSync } from "node:child_process";
+import fs from "node:fs";
+import { createRequire } from "node:module";
+import path from "node:path";
+import { fileURLToPath } from "node:url";
+
+import { assetNameFor } from "./assets.mjs";
+import { download } from "./download.mjs";
+import { HELP, mcpExec, passthroughExec, remoteExec, routeArgv, selectMode } from "./plan.mjs";
+import { resolveBinary } from "./resolve.mjs";
+
+const log = (msg) => console.error(msg);
+
+function pinnedRelease() {
+  const pkgPath = path.join(path.dirname(fileURLToPath(import.meta.url)), "..", "package.json");
+  const pkg = JSON.parse(fs.readFileSync(pkgPath, "utf8"));
+  return { release: pkg.lilbee.release, repo: pkg.lilbee.repo };
+}
+
+function whichSync(name) {
+  try {
+    const tool = process.platform === "win32" ? "where.exe" : "which";
+    const out = execFileSync(tool, [name], { encoding: "utf8", stdio: ["ignore", "pipe", "ignore"] });
+    const first = out.split(/\r?\n/).find(Boolean);
+    return first || null;
+  } catch {
+    return null;
+  }
+}
+
+function mcpRemoteBin() {
+  const require = createRequire(import.meta.url);
+  const pkgJson = require.resolve("mcp-remote/package.json");
+  const pkg = JSON.parse(fs.readFileSync(pkgJson, "utf8"));
+  const rel = typeof pkg.bin === "string" ? pkg.bin : pkg.bin["mcp-remote"];
+  return path.join(path.dirname(pkgJson), rel);
+}
+
+function spawnAndForward({ cmd, args }) {
+  const child = spawn(cmd, args, { stdio: "inherit" });
+  child.on("exit", (code, signal) => {
+    if (signal) process.kill(process.pid, signal);
+    else process.exit(code ?? 0);
+  });
+  child.on("error", (err) => {
+    log(`lilbee: failed to start ${cmd}: ${err.message}`);
+    process.exit(1);
+  });
+  for (const sig of ["SIGINT", "SIGTERM"]) {
+    process.on(sig, () => child.kill(sig));
+  }
+}
+
+async function resolveLocalBinary(env) {
+  const { release, repo } = pinnedRelease();
+  const effectiveRelease = env.LILBEE_RELEASE || release;
+  const assetName = assetNameFor(process.platform, process.arch, env.LILBEE_VARIANT || "");
+  const resolved = await resolveBinary({
+    env: { LILBEE_REPO: repo, ...env },
+    release: effectiveRelease,
+    assetName,
+    deps: { existsSync: fs.existsSync, whichSync, download: (o) => download({ ...o, log }) },
+  });
+  log(`lilbee: using binary from ${resolved.source} (${resolved.path})`);
+  return resolved;
+}
+
+export async function run(argv) {
+  const env = process.env;
+  if (argv[0] === "--launcher-help" || (argv[0] === "--help" && argv.length === 1)) {
+    log(HELP);
+    if (argv[0] === "--launcher-help") return;
+  }
+  const route = routeArgv(argv);
+
+  if (route.kind === "prepare") {
+    const resolved = await resolveLocalBinary(env);
+    log(`lilbee: ready (${resolved.path}).`);
+    return;
+  }
+
+  if (route.kind === "mcp" && selectMode(env) === "remote") {
+    if (!env.LILBEE_TOKEN) {
+      log(
+        "lilbee: LILBEE_URL is set but LILBEE_TOKEN is not — connecting without " +
+          "auth. lilbee servers normally require the session token from server.json."
+      );
+    }
+    log(`lilbee: bridging MCP to ${env.LILBEE_URL}`);
+    spawnAndForward(remoteExec(env, mcpRemoteBin()));
+    return;
+  }
+
+  const resolved = await resolveLocalBinary(env);
+  if (route.kind === "mcp") {
+    spawnAndForward(mcpExec(env, resolved.path, route.args));
+  } else {
+    spawnAndForward(passthroughExec(resolved.path, route.argv));
+  }
+}
+
+export function runAndReport(argv) {
+  run(argv).catch((err) => {
+    log(`lilbee: ${err.message}`);
+    process.exit(1);
+  });
+}
