+++ packages/lilbee/test/assets.test.mjs
